Rewrite method references to intercepted methods in bytecode
The idea is simple: whenever we see an invokedynamic call that has a
MethodHandle of the "interesting" method, we generate a static method
with a compatible signature that calls the interceptor and replace the
method handle with the handle of this generated method. This is similar
to how the actual lambdas are compiled. For example, code like:
```
  callLater(new File("foo")::exists)
```
in a class named `foo.Bar` is rewritten into an equivalent of:
```
  var f = new File("foo");
  callLater(() -> interceptExists(f, "foo.Bar"));
```

One significant difference is the interception of constructors. While
we cannot replace the call to constructor with the call to interceptor
because it will break superclass constructor call, we can freely replace
the constructor reference with the reference to a static method that
calls interceptor.