Allow integ tests to specify java home with Jvm
The Jvm type is a rich wrapper over an (often probed) JVM installation. It contains extra information
about a given java home, including the JVM's version (if probed). This JVM version can prove useful
when executing a Gradle build with that JVM. This commit does not _use_ the actual JVM version exposed
by this change, but makes it available to future work which _does_ use the JVM version.

We introduce a withJvm(Jvm) to replace the existing withJavaHome(File) method. In most cases, we already
had a Jvm object and called withJavaHome(jvm.javaHome). This is now replaced in most places to
withJvm(jvm).

In some cases, we still want to to start the gradle build with a non-standard JVM that does not have a probed
Jvm wrapper. For these cases, we continue to expose a withJavaHome(String) method.

Along the way, we improve the nullability annotations of the Jvm class, and also improve some tests that
test against embedded JREs.